{"schema":"decodex/commit/1","summary":"Split MCP server","authority":"XY-1115"}
